// src/frontend/src/app/modules/fleet-insights/models/fleet-insights-sort-by-types.ts
export enum FleetInsightsSortByTypes {
  AircraftId = 0,
  AircraftSerialNumber = 1,
  AircraftAgeYears = 2,
  AircraftRegistrationNumber = 3,
  Status = 4,
  AircraftManufacturer = 5,
  AircraftSeries = 6,
  EngineSeries = 7,
  AircraftUsage = 8,
  Operator = 9,
  Manager = 10,
  Owner = 11,
  Ownership = 12,
  LeaseStatus = 13,
  LeaseStart = 14,
  LeaseEnd = 15,
  Availability = 16,
  DeliveryDate = 17
}